requirements.txt
fastapi
uvicorn[standard]
pyfluidsynth